// src/app/models/Users.ts
export interface Users {
    id?: number;
    NrInregDGSR: string;
    POD: string;
    NumeClient: string;
    Strada: string;
    Nr: string;
    Bloc: string;
    Scara: string;
    Apartament: string;
    Localitate: string;
    Judet: string;
    Telefon: string;
    E_mail: string;
    NrContractServicii: string;
    DataContractServicii: string;
    DataUltimeiVerificari: string;
    DataScadentaVerifiNotif: string;
}